ipc: search
