+++ src/vs/workbench/services/search/node/rawSearchService.ts
@@ -11,46 +11,10 @@ import gracefulFs = require('graceful-fs');
 gracefulFs.gracefulify(fs);
 
 import {PPromise} from 'vs/base/common/winjs.base';
-import glob = require('vs/base/common/glob');
 import {MAX_FILE_SIZE} from 'vs/platform/files/common/files';
-import {IProgress, ILineMatch, IPatternInfo} from 'vs/platform/search/common/search';
 import {FileWalker, Engine as FileSearchEngine} from 'vs/workbench/services/search/node/fileSearch';
 import {Engine as TextSearchEngine} from 'vs/workbench/services/search/node/textSearch';
-
-export interface IRawSearch {
-	rootFolders: string[];
-	extraFiles?: string[];
-	filePattern?: string;
-	excludePattern?: glob.IExpression;
-	includePattern?: glob.IExpression;
-	contentPattern?: IPatternInfo;
-	maxResults?: number;
-	maxFilesize?: number;
-	fileEncoding?: string;
-}
-
-export interface IRawSearchService {
-	fileSearch(search: IRawSearch): PPromise<ISerializedSearchComplete, ISerializedSearchProgressItem>;
-	textSearch(search: IRawSearch): PPromise<ISerializedSearchComplete, ISerializedSearchProgressItem>;
-}
-
-export interface ISearchEngine {
-	search: (onResult: (match: ISerializedFileMatch) => void, onProgress: (progress: IProgress) => void, done: (error: Error, isLimitHit: boolean) => void) => void;
-	cancel: () => void;
-}
-
-export interface ISerializedSearchComplete {
-	limitHit: boolean;
-}
-
-export interface ISerializedFileMatch {
-	path?: string;
-	lineMatches?: ILineMatch[];
-}
-
-export interface ISerializedSearchProgressItem extends ISerializedFileMatch, IProgress {
-	// Marker interface to indicate the possible values for progress calls from the engine
-}
+import {IRawSearchService, IRawSearch, ISerializedSearchProgressItem, ISerializedSearchComplete, ISearchEngine} from './search';
 
 export class SearchService implements IRawSearchService {
